♻️ refactor: store, review page 분리
