+++ lib/blocs/store/store_bloc.dart
@@ -12,7 +12,6 @@ class StoreBloc extends Bloc<StoreEvent, StoreState> {
     on<StoreEventFetchLocal>(handleFetchLocal);
     on<StoreEventFetchSearch>(handleFetchSearch);
     on<StoreFetchComplete>(_handleStoreFetchComplete);
-    on<StoreEventStoreSelect>(_handleStoreEventStoreSelect);
   }
 
   final StoreRepository storeRepository =
@@ -37,10 +36,5 @@ class StoreBloc extends Bloc<StoreEvent, StoreState> {
   void _handleStoreFetchComplete(
       StoreFetchComplete event, Emitter<StoreState> emit) {
     emit(state.copyWith(fetching: false));
-  }
-
-  void _handleStoreEventStoreSelect(
-      StoreEventStoreSelect event, Emitter<StoreState> emit) {
-    emit(state.copyWith(selectStore: event.selectStore));
   }
 }
