+++ lib/config/routes/routes.dart
@@ -3,6 +3,8 @@ class Routes {
   static const String signIn = '/signIn';
   static const String signUp = '/signUp';
   static const String main = "/main";
+  static const String storeDetail = "/store_detail";
+  static const String review = "/review";
 }
 
 class Tab {
