+++ lib/view/pages/main/home_page.dart
@@ -1,9 +1,7 @@
 import 'package:cakeke/blocs/home/home_bloc.dart';
-import 'package:cakeke/blocs/home/home_event.dart';
 import 'package:cakeke/blocs/home/home_state.dart';
 import 'package:cakeke/view/pages/main/home/home_main_page.dart';
 import 'package:cakeke/view/pages/main/home/home_store_list_page.dart';
-import 'package:cakeke/view/pages/main/map/store_detail_page.dart';
 import 'package:cakeke/view/widgets/common/scaffold_layout.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -23,16 +21,9 @@ class HomePage extends StatelessWidget {
         builder: (context, state) {
           return IndexedStack(
             index: state.selectedPage,
-            children: [
-              const HomeMainPage(),
-              const HomeStoreListPage(),
-              StoreDetailPage(
-                onBackButtonPressed: () {
-                  context
-                      .read<HomeBloc>()
-                      .add(HomePageChanged(selectedPage: state.prevPage));
-                },
-              )
+            children: const [
+              HomeMainPage(),
+              HomeStoreListPage(),
             ],
           );
         },
