+++ lib/view/pages/main/map/map_page.dart
@@ -1,9 +1,6 @@
 import 'package:cakeke/blocs/map/map_bloc.dart';
-import 'package:cakeke/blocs/map/map_event.dart';
 import 'package:cakeke/blocs/map/map_state.dart';
 import 'package:cakeke/view/pages/main/map/map_main_page.dart';
-import 'package:cakeke/view/pages/main/map/review_page.dart';
-import 'package:cakeke/view/pages/main/map/store_detail_page.dart';
 import 'package:cakeke/view/pages/main/map/store_list_page.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter/material.dart';
@@ -17,18 +14,7 @@ class MapPage extends StatelessWidget {
       builder: (context, state) {
         return IndexedStack(
           index: state.selectedPage,
-          children: [
-            const MapMainPage(),
-            const StoreListPage(),
-            StoreDetailPage(
-              onBackButtonPressed: () {
-                context
-                    .read<MapBloc>()
-                    .add(const MapPageChanged(selectedPage: 0));
-              },
-            ),
-            const ReviewPage()
-          ],
+          children: const [MapMainPage(), StoreListPage()],
         );
       },
     );
